requirements.txt
python-dotenv
pydantic
pydantic-settings
langchain
langchain-community
langgraph
groq
sentence-transformers
chromadb
pypdf
pandas
faker
streamlit
requests
tavily-python
fastapi
uvicorn
loguru


